Split Home child components
